+++ .storybook/main.ts
@@ -21,6 +21,7 @@ const config: StorybookConfig = {
 	staticDirs: ['../example-site/public'],
 	stories: [
 		'../packages/*/src/**/*.stories.@(ts|tsx)',
+		'../example-site/**/*.stories.@(ts|tsx)',
 		'./stories/*.@(ts|tsx)',
 		'./stories/*/index.@(ts|tsx)',
 	],
